Extract ether-router to own file
