+++ lib/proxyResolver.js
@@ -1,4 +1,4 @@
-const ethers = require("ethers");
+const etherRouter = require("./etherRouter");
 const SyncRequest = require("./syncRequest");
 
 class ProxyResolver {
@@ -10,7 +10,7 @@ class ProxyResolver {
     if (typeof config.proxyResolver === "function") {
       this.resolve = config.proxyResolver.bind(this);
     } else if (config.proxyResolver === "EtherRouter") {
-      this.resolve = this._etherRouter;
+      this.resolve = etherRouter.bind(this);
     } else if (config.proxyResolver === "ZOS@2") {
       this.resolve = this._zos2;
     } else {
@@ -45,55 +45,6 @@ class ProxyResolver {
     if (match) {
       this.data.trackNameByAddress(match.name, address);
       return match.name;
-    }
-  }
-
-  /**
-   * Queries EtherRouter for address of resolver.
-   * Queries Resolver for contract address w/ method signature.
-   * Returns contract name matching the resolved address.
-   * @param  {Object} transaction result of web3.eth.getTransaction
-   * @return {String}             contract name
-   */
-  _etherRouter(transaction) {
-    let contractAddress;
-    let contractName;
-
-    try {
-      const ABI = ["function resolver()", "function lookup(bytes4 sig)"];
-      const iface = new ethers.utils.Interface(ABI);
-      const signature = transaction.input.slice(0, 10);
-
-      // EtherRouter.resolver
-      const resolverAddress = this.sync.call(
-        {
-          to: transaction.to,
-          data: iface.functions.resolver.encode([])
-        },
-        transaction.blockNumber
-      );
-
-      // Resolver.lookup(sig)
-      contractAddress = this.sync.call(
-        {
-          to: ethers.utils.hexStripZeros(resolverAddress),
-          data: iface.functions.lookup.encode([signature])
-        },
-        transaction.blockNumber
-      );
-    } catch (err) {
-      this.unresolvedCalls++;
-      return;
-    }
-
-    // Geth returns `undefined` for Truffle's solidity tests here.
-    if (contractAddress) {
-      contractAddress = ethers.utils.hexStripZeros(contractAddress);
-      contractName = this.data.getNameByAddress(contractAddress);
-
-      // Try to resolve by deployedBytecode
-      if (contractName) return contractName;
-      else return this.resolveByDeployedBytecode(contractAddress);
     }
   }
 }
